pyproject.toml
[tool.poetry]
name = "sastocks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
sastocks = "sastocks.main:app"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.0.350"
nltk = "^3.8.1"
SQLAlchemy = "^2.0.23"
click = "^8.1.7"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
yfinance = "^0.2.33"
typer = "^0.9.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
mentat = "^1.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
